IsGroupSelect used to optionally enable UI lock per player. Move files into UiRebinding package.
