+++ Packages/UiRebinding/UiMultiSelect.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4b9130325d33ec640917b60f1cc64466
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
